fix exception handling, refactor tests
